Add version command for all binaries/symlinks
